fix(#111): break transcript-file-state import cycle for check-guards
Extract stripStaleThinkingBlocksFromSessionManagerBranch and
rewriteTranscriptEntriesInSessionManager into a new
strip-stale-thinking-blocks.ts module so session-manager.ts no longer
imports transcript-rewrite.ts at runtime.

Root cause: session-manager.ts imported
stripStaleThinkingBlocksFromSessionManagerBranch (a runtime value) from
transcript-rewrite.ts, which imports transcript-file-state.ts, which
imported buildSessionContext/migrateSessionEntries/parseSessionEntries
via the sessions/index.ts barrel (which re-exports session-manager.ts).
That closed a runtime import cycle flagged by pnpm check:import-cycles.

Fix (smallest behavior-preserving change):
- New module embedded-agent-runner/strip-stale-thinking-blocks.ts holds
  stripStaleThinkingBlocksFromSessionManagerBranch and
  rewriteTranscriptEntriesInSessionManager (plus their private helpers),
  with no dependency on transcript-file-state.ts.
- session-manager.ts now imports the strip function from the new module
  instead of transcript-rewrite.ts.
- transcript-file-state.ts now imports buildSessionContext /
  migrateSessionEntries / parseSessionEntries directly from
  session-manager.ts instead of the sessions/index.ts barrel.
- transcript-rewrite.ts keeps its own copies of the small pure helpers
  (estimateMessageBytes, findTranscriptRewriteMatches, remapEntryId,
  SessionManagerLike/SessionBranchEntry types) needed by the
  file/state-based rewrite path (rewriteTranscriptEntriesInState),
  since importing them back from strip-stale-thinking-blocks.ts would
  recreate the cycle.
- Updated all production importers (context-engine-maintenance.ts,
  thinking-replay-repair.ts, tool-result-truncation.ts) and test mocks/
  spies to import rewriteTranscriptEntriesInSessionManager from the new
  module.

No behavior change: all moved code is copied verbatim, only import
paths changed. Does not touch rotateTranscriptAfterCompaction or
buildSuccessorEntries (C-class item #5, untouched).

Verified: pnpm check:import-cycles now reports 0 runtime value cycles
(was 1, 14-node component). pnpm tsgo:prod passes. Per-file vitest runs
green: transcript-rewrite.test.ts (6/6), context-engine-maintenance.test.ts
(21/21), agent-session-thinking-cleanup.test.ts (17/17),
transcript-file-state.test.ts (25/25), tool-result-truncation.test.ts
(38/38), thinking-replay-repair.test.ts (4/4). oxlint clean on all
touched files.
